+++ client/src/components/PageTitle/style.ts
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const Title = styled.header`
+display: flex;
+flex: row wrap;
+justify-content: flex-start;
+align-items: center;
+gap: 8px;
+height: 48px;
+
+span {
+  font-family: "Gothic-M";
+  font-size: 32px;
+}
+`;
